refactor(daytona): consolidate integration modules into cohesive targets
Merge 11 thin-wrapper files into 5 consolidated modules, reducing the
Daytona integration surface from 22 to 14 active files.

Consolidations:
- config.py ← absorbed client.py (SDK error classification)
- types.py ← absorbed payload_models.py, spec_runtime.py, state_helpers.py
- volume_runtime.py ← absorbed volume_browser.py, volume_inventory.py, volume_mounts.py
- workspace_runtime.py ← absorbed workspace_paths.py, repo_refs.py, repo_checkout.py
- session_runtime.py ← absorbed admin_runtime.py

Dead code removed:
- filesystem.py (unused since SDK native fs)
- runtime_support.py (superseded by async_compat + config)

Documentation:
- Updated src/fleet_rlm/AGENTS.md with consolidated module inventory
- Added TODO(sdk-sync), TODO(sdk-enum), and FUTURE comments for
  SDK workarounds that can be removed when upstream ships fixes

All monkeypatch test paths updated to reference new module locations.
Validated: ruff, ty, and targeted test suites pass.
