+++ src/fleet_rlm/integrations/daytona/child_isolation.py
@@ -9,7 +9,8 @@ from typing import Any, Literal, cast
 from fleet_rlm.runtime.execution.interpreter_support import initialize_sub_rlm_state
 from fleet_rlm.runtime.execution.profiles import ExecutionProfile
 
-from .runtime import DaytonaSandboxRuntime, DaytonaSandboxSession
+from .runtime import DaytonaSandboxRuntime
+from .session_runtime import DaytonaSandboxSession
 from .types import SandboxSpec
 
 ChildIsolationMode = Literal["auto", "context"]
