--- src/fleet_rlm/integrations/daytona/runtime_support.py
@@ -1,216 +1,0 @@
-"""Low-level Daytona runtime support helpers."""
-
-from __future__ import annotations
-
-import asyncio
-import logging
-import time as _time
-from pathlib import PurePosixPath
-from typing import Any
-
-from fleet_rlm.runtime.execution.storage_paths import mounted_storage_roots
-
-from .async_compat import _await_if_needed
-from .config import ResolvedDaytonaConfig
-from .diagnostics import DaytonaDiagnosticError, VolumeNotReadyError
-
-logger = logging.getLogger(__name__)
-
-DAYTONA_PERSISTENT_VOLUME_MOUNT_PATH = PurePosixPath("/home/daytona/memory")
-
-_REMOTE_DIRECTORY_MODE = "755"
-_VOLUME_READY_STATES = frozenset({"ready"})
-_VOLUME_ERROR_STATES = frozenset({"error", "failed", "deleted"})
-
-
-def daytona_import_error(exc: ImportError) -> RuntimeError:
-    """Build the standard Daytona SDK missing-dependency error."""
-    return RuntimeError(
-        "Daytona SDK is not available. Install dependencies with `uv sync` "
-        "and configure DAYTONA_API_KEY / DAYTONA_API_URL before using Daytona "
-        "commands. See https://www.daytona.io/docs/en/python-sdk/"
-    )
-
-
-def build_daytona_client(config: ResolvedDaytonaConfig) -> Any:
-    """Build an AsyncDaytona client lazily to keep imports light."""
-    try:
-        from daytona import AsyncDaytona, DaytonaConfig
-    except ImportError as exc:  # pragma: no cover - environment specific
-        raise daytona_import_error(exc) from exc
-    return AsyncDaytona(
-        DaytonaConfig(
-            api_key=config.api_key,
-            api_url=config.api_url.rstrip("/"),
-            target=config.target,
-        )
-    )
-
-
-async def aensure_remote_directory(fs: Any, remote_path: PurePosixPath) -> None:
-    """Ensure a remote Daytona directory exists."""
-    directory = str(remote_path)
-    if directory and directory not in {".", "/"}:
-        await _await_if_needed(fs.create_folder(directory, _REMOTE_DIRECTORY_MODE))
-
-
-def canonicalize_volume_state_token(value: Any) -> str:
-    """Normalize raw Daytona SDK volume state values."""
-    candidates: list[str] = []
-
-    if isinstance(value, str):
-        candidates.append(value)
-    else:
-        state_value = getattr(value, "value", None)
-        if state_value not in (None, ""):
-            candidates.append(str(state_value))
-        state_name = getattr(value, "name", None)
-        if state_name not in (None, ""):
-            candidates.append(str(state_name))
-        if value not in (None, ""):
-            candidates.append(str(value))
-
-    for candidate in candidates:
-        normalized = candidate.strip().lower()
-        if not normalized:
-            continue
-        normalized = normalized.replace("-", "_").replace(" ", "_")
-        if "." in normalized:
-            normalized = normalized.rsplit(".", 1)[-1]
-        if normalized:
-            return normalized
-    return ""
-
-
-def volume_state_details(volume: Any) -> tuple[str, str]:
-    """Return raw and normalized SDK volume state strings."""
-    raw_state_value = getattr(volume, "state", None)
-    raw_state = str(raw_state_value or "").strip()
-    normalized_state = canonicalize_volume_state_token(raw_state_value)
-    return raw_state, normalized_state
-
-
-def volume_state_missing(volume: Any, *, raw_state: str, normalized_state: str) -> bool:
-    """Return whether the SDK response omitted a usable state token."""
-    if raw_state or normalized_state:
-        return False
-    return bool(getattr(volume, "id", None))
-
-
-def raise_if_volume_error(
-    volume_name: str,
-    *,
-    raw_state: str,
-    normalized_state: str,
-) -> None:
-    """Raise diagnostics when a volume is in a terminal error state."""
-    if normalized_state in _VOLUME_ERROR_STATES:
-        message = f"Volume '{volume_name}' is in error state '{normalized_state}'"
-        if raw_state and raw_state != normalized_state:
-            message = (
-                f"Volume '{volume_name}' is in error state "
-                f"'{normalized_state}' (raw='{raw_state}')"
-            )
-        raise DaytonaDiagnosticError(
-            message,
-            category="sandbox_create_clone_error",
-            phase="sandbox_create",
-        )
-
-
-async def await_volume_ready(
-    client: Any,
-    volume_name: str,
-    volume: Any,
-    *,
-    timeout: float = 60.0,
-) -> Any:
-    """Poll until a Daytona volume reaches ready state."""
-    raw_state, state = volume_state_details(volume)
-
-    if volume_state_missing(volume, raw_state=raw_state, normalized_state=state):
-        return volume
-    if state in _VOLUME_READY_STATES:
-        return volume
-    raise_if_volume_error(
-        volume_name,
-        raw_state=raw_state,
-        normalized_state=state,
-    )
-
-    deadline = _time.monotonic() + timeout
-    interval = 1.0
-
-    while _time.monotonic() < deadline:
-        logger.debug(
-            (
-                "Volume '%s' not ready "
-                "(raw_state=%s, normalized_state=%s, state_type=%s, state_repr=%r), "
-                "polling in %.1fs"
-            ),
-            volume_name,
-            raw_state or "<empty>",
-            state or "<empty>",
-            type(getattr(volume, "state", None)).__name__,
-            getattr(volume, "state", None),
-            interval,
-        )
-        await asyncio.sleep(interval)
-        interval = min(interval * 2, 10.0)
-
-        volume = await _await_if_needed(client.volume.get(volume_name))
-        raw_state, state = volume_state_details(volume)
-
-        if volume_state_missing(volume, raw_state=raw_state, normalized_state=state):
-            return volume
-        if state in _VOLUME_READY_STATES:
-            return volume
-        raise_if_volume_error(
-            volume_name,
-            raw_state=raw_state,
-            normalized_state=state,
-        )
-
-    raise VolumeNotReadyError(
-        volume_name=volume_name,
-        volume_state=state or raw_state or "unknown",
-        raw_volume_state=raw_state or None,
-        timeout_seconds=timeout,
-    )
-
-
-async def aensure_daytona_volume_layout(
-    *,
-    sandbox: Any,
-    mounted_root: str = str(DAYTONA_PERSISTENT_VOLUME_MOUNT_PATH),
-) -> None:
-    """Ensure canonical durable directories exist on a mounted Daytona volume."""
-    roots = mounted_storage_roots(mounted_root)
-    try:
-        for path in (
-            roots.memory_root,
-            roots.artifacts_root,
-            roots.buffers_root,
-            roots.meta_root,
-        ):
-            await aensure_remote_directory(sandbox.fs, PurePosixPath(path))
-    except Exception as exc:
-        raise DaytonaDiagnosticError(
-            f"Daytona volume layout create failure: {exc}",
-            category="sandbox_create_clone_error",
-            phase="sandbox_create",
-        ) from exc
-
-
-__all__ = [
-    "DAYTONA_PERSISTENT_VOLUME_MOUNT_PATH",
-    "aensure_daytona_volume_layout",
-    "aensure_remote_directory",
-    "await_volume_ready",
-    "build_daytona_client",
-    "canonicalize_volume_state_token",
-    "daytona_import_error",
-    "raise_if_volume_error",
-    "volume_state_details",
-    "volume_state_missing",
-]
