+++ src/fleet_rlm/integrations/daytona/session_runtime.py
@@ -1,0 +1,313 @@
+"""Daytona sandbox session model and session-local helpers."""
+
+from __future__ import annotations
+
+import asyncio
+import threading
+import time
+from contextlib import suppress
+from dataclasses import dataclass, field
+from pathlib import PurePosixPath
+from typing import Any
+
+from .async_compat import _await_if_needed, _run_async_compat
+from .diagnostics import DaytonaDiagnosticError
+from .sandbox_lifecycle import (
+    aarchive_sandbox_session as _aarchive_sandbox_session,
+    adelete_sandbox_session as _adelete_sandbox_session,
+    arecover_sandbox_session as _arecover_sandbox_session,
+    arefresh_sandbox_session_activity as _arefresh_sandbox_session_activity,
+    aresize_sandbox_session as _aresize_sandbox_session,
+    create_lsp_server as _create_lsp_server,
+)
+from .types import ContextSource
+from .volume_runtime import DAYTONA_PERSISTENT_VOLUME_MOUNT_PATH
+
+
+# ---------------------------------------------------------------------------
+# Admin code-execution helpers (formerly admin_runtime.py)
+# ---------------------------------------------------------------------------
+async def _arun_admin_code(
+    *,
+    sandbox: Any,
+    code: str,
+    phase: str,
+    error_prefix: str,
+    category: str = "sandbox_create_clone_error",
+) -> str:
+    """Run administrative code inside a sandbox via ``sandbox.process.code_run``."""
+    try:
+        from daytona.common.process import CodeRunParams
+
+        result = await _await_if_needed(
+            sandbox.process.code_run(code, params=CodeRunParams())
+        )
+    except Exception as exc:
+        raise DaytonaDiagnosticError(
+            f"{error_prefix}: {exc}",
+            category=category,
+            phase=phase,
+        ) from exc
+
+    exit_code = getattr(result, "exit_code", 0)
+    if exit_code:
+        detail = str(
+            getattr(result, "stderr", "")
+            or getattr(result, "result", "")
+            or getattr(getattr(result, "artifacts", None), "stdout", "")
+            or getattr(result, "output", "")
+            or f"process exited with status {exit_code}"
+        )
+        raise DaytonaDiagnosticError(
+            f"{error_prefix}: {detail}",
+            category=category,
+            phase=phase,
+        )
+
+    return str(
+        getattr(result, "stdout", "")
+        or getattr(result, "result", "")
+        or getattr(getattr(result, "artifacts", None), "stdout", "")
+        or getattr(result, "output", "")
+        or ""
+    )
+
+
+# ---------------------------------------------------------------------------
+# Session model
+# ---------------------------------------------------------------------------
+def _current_async_owner() -> tuple[int, int]:
+    return (threading.get_ident(), id(asyncio.get_running_loop()))
+
+
+@dataclass(slots=True)
+class DaytonaSandboxSession:
+    """Concrete Daytona workspace session backed by a sandbox and interpreter context."""
+
+    sandbox: Any
+    repo_url: str | None
+    ref: str | None
+    volume_name: str | None
+    workspace_path: str
+    context_sources: list[ContextSource] = field(default_factory=list)
+    phase_timings_ms: dict[str, int] = field(default_factory=dict)
+    volume_mount_path: str = str(DAYTONA_PERSISTENT_VOLUME_MOUNT_PATH)
+    context_id: str | None = None
+    owner_thread_id: int | None = None
+    owner_loop_id: int | None = None
+    execution_event_callback: Any | None = None
+    _context: Any | None = field(default=None, init=False, repr=False)
+    _driver_started: bool = field(default=False, init=False, repr=False)
+    _runtime_ref: Any | None = field(default=None, init=False, repr=False)
+
+    @property
+    def sandbox_id(self) -> str | None:
+        return str(getattr(self.sandbox, "id", "") or "") or None
+
+    def bind_current_async_owner(self) -> None:
+        self.owner_thread_id, self.owner_loop_id = _current_async_owner()
+
+    def matches_current_async_owner(self) -> bool:
+        if self.owner_thread_id is None or self.owner_loop_id is None:
+            return False
+        try:
+            return (self.owner_thread_id, self.owner_loop_id) == _current_async_owner()
+        except RuntimeError:
+            return False
+
+    async def aensure_context(self) -> Any:
+        if self._context is not None:
+            return self._context
+        if not self.matches_current_async_owner() and self._runtime_ref is not None:
+            sandbox_id = self.sandbox_id
+            if sandbox_id:
+                with suppress(Exception):
+                    self.sandbox = await self._runtime_ref._aget_sandbox(
+                        sandbox_id,
+                        recover=False,
+                    )
+                    self.bind_current_async_owner()
+        if self.context_id:
+            existing_contexts: list[Any] | None = None
+            with suppress(Exception):
+                existing_contexts = await _await_if_needed(
+                    self.sandbox.code_interpreter.list_contexts()
+                )
+            if existing_contexts is not None:
+                for existing in existing_contexts:
+                    if str(getattr(existing, "id", "") or "") == self.context_id:
+                        self._context = existing
+                        return existing
+        context = await _await_if_needed(
+            self.sandbox.code_interpreter.create_context(cwd=self.workspace_path)
+        )
+        self._context = context
+        self.context_id = str(getattr(context, "id", "") or "") or None
+        return context
+
+    def ensure_context(self) -> Any:
+        return _run_async_compat(self.aensure_context)
+
+    async def astart_driver(self, *, timeout: float = 30.0) -> None:
+        _ = timeout
+        await self.aensure_context()
+        self._driver_started = True
+
+    def start_driver(self, *, timeout: float = 30.0) -> None:
+        _run_async_compat(self.astart_driver, timeout=timeout)
+
+    async def aclose_driver(self) -> None:
+        self._driver_started = False
+
+    def close_driver(self) -> None:
+        _run_async_compat(self.aclose_driver)
+
+    async def adelete_context(self) -> None:
+        context = self._context
+        self._context = None
+        if context is None and self.context_id:
+            with suppress(Exception):
+                existing_contexts = await _await_if_needed(
+                    self.sandbox.code_interpreter.list_contexts()
+                )
+                for existing in existing_contexts:
+                    if str(getattr(existing, "id", "") or "") == self.context_id:
+                        context = existing
+                        break
+        if context is not None:
+            with suppress(Exception):
+                await _await_if_needed(
+                    self.sandbox.code_interpreter.delete_context(context)
+                )
+        self.context_id = None
+        self._driver_started = False
+
+    def delete_context(self) -> None:
+        _run_async_compat(self.adelete_context)
+
+    def _resolve_sandbox_path(self, path: str) -> str:
+        candidate = PurePosixPath(str(path or "").strip() or ".")
+        if candidate.is_absolute():
+            return str(candidate)
+        return str(PurePosixPath(self.workspace_path) / candidate)
+
+    async def _arebind_sandbox_if_needed(self) -> None:
+        if self.matches_current_async_owner() or self._runtime_ref is None:
+            return
+        sandbox_id = self.sandbox_id
+        if not sandbox_id:
+            return
+        with suppress(Exception):
+            self.sandbox = await self._runtime_ref._aget_sandbox(
+                sandbox_id,
+                recover=False,
+            )
+            self.bind_current_async_owner()
+
+    async def aread_file(self, path: str) -> str:
+        await self._arebind_sandbox_if_needed()
+        raw = await _await_if_needed(
+            self.sandbox.fs.download_file(self._resolve_sandbox_path(path))
+        )
+        if raw is None:
+            return ""
+        if isinstance(raw, str):
+            return raw
+        return bytes(raw).decode("utf-8", errors="replace")
+
+    def read_file(self, path: str) -> str:
+        return _run_async_compat(self.aread_file, path)
+
+    async def awrite_file(self, path: str, content: str) -> str:
+        await self._arebind_sandbox_if_needed()
+        resolved_path = self._resolve_sandbox_path(path)
+        payload = content.encode("utf-8")
+        callback = getattr(self, "execution_event_callback", None)
+        if callable(callback):
+            callback(
+                {
+                    "phase": "progress",
+                    "timestamp": time.time(),
+                    "execution_profile": "durable_write",
+                    "code_hash": "durable-write",
+                    "code_preview": "sandbox.fs.upload_file",
+                    "event_kind": "durable_write_started",
+                    "path": resolved_path,
+                    "bytes_total": len(payload),
+                    "bytes_written": 0,
+                }
+            )
+        await _await_if_needed(self.sandbox.fs.upload_file(payload, resolved_path))
+        if callable(callback):
+            callback(
+                {
+                    "phase": "progress",
+                    "timestamp": time.time(),
+                    "execution_profile": "durable_write",
+                    "code_hash": "durable-write",
+                    "code_preview": "sandbox.fs.upload_file",
+                    "event_kind": "durable_write_completed",
+                    "path": resolved_path,
+                    "bytes_total": len(payload),
+                    "bytes_written": len(payload),
+                }
+            )
+        return resolved_path
+
+    def write_file(self, path: str, content: str) -> str:
+        return _run_async_compat(self.awrite_file, path, content)
+
+    async def alist_files(self, path: str) -> list[Any]:
+        await self._arebind_sandbox_if_needed()
+        entries = await _await_if_needed(
+            self.sandbox.fs.list_files(self._resolve_sandbox_path(path))
+        )
+        return list(entries)
+
+    def list_files(self, path: str) -> list[Any]:
+        return _run_async_compat(self.alist_files, path)
+
+    async def adelete(self) -> None:
+        await _adelete_sandbox_session(self)
+
+    def delete(self) -> None:
+        _run_async_compat(self.adelete)
+
+    async def aarchive(self) -> None:
+        await _aarchive_sandbox_session(self)
+
+    def archive(self) -> None:
+        _run_async_compat(self.aarchive)
+
+    async def arecover(self, *, timeout: float = 60.0) -> None:
+        await _arecover_sandbox_session(self, timeout=timeout)
+
+    def recover(self, *, timeout: float = 60.0) -> None:
+        _run_async_compat(self.arecover, timeout=timeout)
+
+    async def arefresh_activity(self) -> None:
+        await _arefresh_sandbox_session_activity(self)
+
+    def refresh_activity(self) -> None:
+        _run_async_compat(self.arefresh_activity)
+
+    async def aresize(self, *, cpu: int, memory: int, disk: int) -> None:
+        await _aresize_sandbox_session(self, cpu=cpu, memory=memory, disk=disk)
+
+    def resize(self, *, cpu: int, memory: int, disk: int) -> None:
+        _run_async_compat(self.aresize, cpu=cpu, memory=memory, disk=disk)
+
+    def create_lsp_server(
+        self,
+        *,
+        language: str = "python",
+        project_path: str | None = None,
+    ) -> Any:
+        return _create_lsp_server(
+            self,
+            language=language,
+            project_path=project_path,
+        )
+
+
+__all__ = ["DaytonaSandboxSession", "_arun_admin_code"]
