+++ src/fleet_rlm/integrations/daytona/types.py
@@ -1,20 +1,19 @@
-"""Consolidated Daytona type definitions.
-
-Sections (in dependency order):
-1. Serialization helpers
-2. Sandbox spec and LM bootstrap types
-3. Staged-context types
-"""
+"""Core Daytona transport types, payload models, and spec-building helpers."""
 
 from __future__ import annotations
 
+import datetime
 import json
 import re
 from dataclasses import asdict, dataclass, field
-from typing import Any
+from typing import Any, Mapping
+
+from pydantic import BaseModel, Field, field_validator
+
+from fleet_rlm.utils.paths import dedupe_paths
 
 # ---------------------------------------------------------------------------
-# Section 1: Serialization helpers
+# Payload normalization helpers (formerly payload_models.py)
 # ---------------------------------------------------------------------------
 
 _WHITESPACE_RE = re.compile(r"\s+")
@@ -32,28 +31,221 @@ def _normalize_optional_text(value: Any, *, limit: int | None = None) -> str | N
     return collapsed
 
 
-def _coerce_positive_int(value: Any) -> int | None:
-    if value is None or value == "":
-        return None
-    try:
-        parsed = int(value)
-    except (TypeError, ValueError):
-        return None
-    return parsed if parsed > 0 else None
+# ---------------------------------------------------------------------------
+# Pydantic payload models (formerly payload_models.py)
+# ---------------------------------------------------------------------------
 
 
-def _coerce_nonnegative_int(value: Any) -> int | None:
-    if value is None or value == "":
-        return None
-    try:
-        parsed = int(value)
-    except (TypeError, ValueError):
-        return None
-    return parsed if parsed >= 0 else None
+class SandboxLmRuntimeConfig(BaseModel):
+    """Serializable LM bootstrap config passed into sandbox-local runtimes."""
+
+    model: str
+    api_key: str
+    api_base: str | None = None
+    max_tokens: int = 64_000
+    delegate_model: str | None = None
+    delegate_api_key: str | None = None
+    delegate_api_base: str | None = None
+
+    @field_validator("model", "api_key", mode="before")
+    @classmethod
+    def _normalize_required_text(cls, value: Any) -> str:
+        text = _normalize_optional_text(value)
+        if text is None:
+            raise ValueError("Sandbox LM config requires model and api_key.")
+        return text
+
+    @field_validator(
+        "api_base",
+        "delegate_model",
+        "delegate_api_key",
+        "delegate_api_base",
+        mode="before",
+    )
+    @classmethod
+    def _normalize_optional_text_fields(cls, value: Any) -> str | None:
+        return _normalize_optional_text(value)
+
+    @field_validator("max_tokens", mode="before")
+    @classmethod
+    def _normalize_max_tokens(cls, value: Any) -> int:
+        if value is None or value == "":
+            return 64_000
+        try:
+            parsed = int(value)
+        except (TypeError, ValueError):
+            return 64_000
+        return parsed if parsed > 0 else 64_000
+
+    def to_dict(self) -> dict[str, Any]:
+        return self.model_dump(mode="python")
+
+    @classmethod
+    def from_raw(cls, raw: Any) -> SandboxLmRuntimeConfig:
+        if not isinstance(raw, dict):
+            raise ValueError("Sandbox LM config must be a dict.")
+        try:
+            return cls.model_validate(raw)
+        except Exception as exc:  # pragma: no cover - pydantic internals
+            raise ValueError("Sandbox LM config requires model and api_key.") from exc
+
+
+class ContextSource(BaseModel):
+    """Host-sourced local context staged into the Daytona workspace."""
+
+    source_id: str
+    kind: str
+    host_path: str
+    staged_path: str
+    source_type: str | None = None
+    extraction_method: str | None = None
+    file_count: int = 1
+    skipped_count: int = 0
+    warnings: list[str] = Field(default_factory=list)
+
+    @field_validator("source_id", "kind", "host_path", "staged_path", mode="before")
+    @classmethod
+    def _normalize_required_fields(cls, value: Any) -> str:
+        text = _normalize_optional_text(value)
+        if text is None:
+            raise ValueError(
+                "Context source requires source_id, kind, host_path, and staged_path."
+            )
+        return text
+
+    @field_validator("source_type", "extraction_method", mode="before")
+    @classmethod
+    def _normalize_optional_fields(cls, value: Any) -> str | None:
+        return _normalize_optional_text(value)
+
+    @field_validator("file_count", mode="before")
+    @classmethod
+    def _normalize_file_count(cls, value: Any) -> int:
+        if value is None or value == "":
+            return 1
+        try:
+            parsed = int(value)
+        except (TypeError, ValueError):
+            return 1
+        return parsed if parsed > 0 else 1
+
+    @field_validator("skipped_count", mode="before")
+    @classmethod
+    def _normalize_skipped_count(cls, value: Any) -> int:
+        if value is None or value == "":
+            return 0
+        try:
+            parsed = int(value)
+        except (TypeError, ValueError):
+            return 0
+        return parsed if parsed >= 0 else 0
+
+    @field_validator("warnings", mode="before")
+    @classmethod
+    def _normalize_warnings(cls, value: Any) -> list[str]:
+        if value is None or value == "":
+            return []
+        if isinstance(value, (str, bytes)):
+            items = [value]
+        else:
+            try:
+                items = list(value)
+            except TypeError:
+                items = [value]
+        normalized: list[str] = []
+        for item in items:
+            text = _normalize_optional_text(item)
+            if text is not None:
+                normalized.append(text)
+        return normalized
+
+    def to_dict(self) -> dict[str, Any]:
+        return self.model_dump(mode="python")
+
+    @classmethod
+    def from_raw(cls, raw: Any) -> ContextSource:
+        if not isinstance(raw, dict):
+            raise ValueError("Context source payload must be a dict.")
+        try:
+            return cls.model_validate(raw)
+        except Exception as exc:  # pragma: no cover - pydantic internals
+            raise ValueError(
+                "Context source requires source_id, kind, host_path, and staged_path."
+            ) from exc
 
 
 # ---------------------------------------------------------------------------
-# Section 2: Sandbox spec and LM bootstrap types
+# State normalization helpers (formerly state_helpers.py)
+# ---------------------------------------------------------------------------
+
+
+def render_final_text(value: Any) -> str:
+    """Extract and normalize final text from structured output."""
+    if value is None:
+        return ""
+    if isinstance(value, str):
+        return value
+    if isinstance(value, dict):
+        for key in ("final_markdown", "summary", "text", "content", "message"):
+            candidate = value.get(key)
+            if isinstance(candidate, str) and candidate.strip():
+                return candidate
+        nested_value = value.get("value")
+        if nested_value is not value:
+            nested_text = render_final_text(nested_value)
+            if nested_text:
+                return nested_text
+    try:
+        return json.dumps(value, indent=2, ensure_ascii=False)
+    except (TypeError, ValueError):
+        return str(value)
+
+
+def history_messages(history: Any) -> list[dict[str, str]]:
+    """Extract message list from a history object."""
+    messages = getattr(history, "messages", [])
+    if isinstance(messages, list):
+        return [item for item in messages if isinstance(item, dict)]
+    return []
+
+
+def normalize_history_turn(raw: dict[str, Any]) -> dict[str, str] | None:
+    """Normalize a single history turn into ``{user_request, assistant_response}``."""
+    user_request = str(raw.get("user_request", "") or "").strip()
+    assistant_response = render_final_text(raw.get("assistant_response", "")).strip()
+    if not user_request and not assistant_response:
+        return None
+    return {
+        "user_request": user_request,
+        "assistant_response": assistant_response,
+    }
+
+
+def normalized_history_messages(history: Any) -> list[dict[str, str]]:
+    """Return a clean list of normalized history turns."""
+    normalized: list[dict[str, str]] = []
+    for item in history_messages(history):
+        turn = normalize_history_turn(item)
+        if turn is not None:
+            normalized.append(turn)
+    return normalized
+
+
+def normalized_context_sources(raw: Any) -> list[ContextSource]:
+    """Normalize a raw sources list into validated ``ContextSource`` objects."""
+    if not isinstance(raw, list):
+        return []
+    normalized: list[ContextSource] = []
+    for item in raw:
+        try:
+            normalized.append(ContextSource.from_raw(item))
+        except Exception:
+            continue
+    return normalized
+
+
+# ---------------------------------------------------------------------------
+# Exceptions and core types
 # ---------------------------------------------------------------------------
 
 
@@ -90,8 +282,8 @@ class SandboxSpec:
     env_vars: dict[str, str] | None = None
     labels: dict[str, str] | None = None
     ephemeral: bool = True
-    auto_stop_interval: int | None = 30  # provider minutes; refresh_activity() resets
-    auto_archive_interval: int | None = 60  # provider minutes
+    auto_stop_interval: int | None = 30
+    auto_archive_interval: int | None = 60
     auto_delete_interval: int | None = None
     cpu: int | None = None
     memory: int | None = None
@@ -120,13 +312,13 @@ class SandboxSpec:
             params["snapshot"] = self.snapshot
         if self.cpu is not None or self.memory is not None or self.disk is not None:
             params["resources"] = {
-                k: v
-                for k, v in [
+                key: value
+                for key, value in [
                     ("cpu", self.cpu),
                     ("memory", self.memory),
                     ("disk", self.disk),
                 ]
-                if v is not None
+                if value is not None
             }
         if self.network_block_all is not None:
             params["network_block_all"] = self.network_block_all
@@ -143,12 +335,7 @@ class SandboxSpec:
         return params
 
     def _daytona_lifecycle_params(self) -> dict[str, int]:
-        """Return Daytona lifecycle settings in provider-minute units.
-
-        Keep the minute-unit shaping in one Daytona-specific helper so the
-        SDK boundary stays explicit even though the dataclass fields retain the
-        provider's original parameter names.
-        """
+        """Return Daytona lifecycle settings in provider-minute units."""
         params: dict[str, int] = {}
         if self.auto_stop_interval is not None:
             params["auto_stop_interval"] = self.auto_stop_interval
@@ -159,106 +346,11 @@ class SandboxSpec:
         return params
 
     def to_create_params(self, *, volume_id: str | None = None) -> dict[str, Any]:
-        """Build keyword arguments for the SDK create-params constructor.
-
-        When ``image`` is set the returned dict includes an ``"image"`` key
-        carrying the ``daytona.Image`` object (for ``CreateSandboxFromImageParams``).
-        Otherwise the dict is suitable for ``CreateSandboxFromSnapshotParams``.
-        """
+        """Build keyword arguments for the SDK create-params constructor."""
         params = self._common_params(volume_id=volume_id)
         if self.image is not None:
             params["image"] = self.image
         return params
-
-
-@dataclass(slots=True)
-class SandboxLmRuntimeConfig:
-    """Serializable LM bootstrap config passed into sandbox-local runtimes."""
-
-    model: str
-    api_key: str
-    api_base: str | None = None
-    max_tokens: int = 64_000
-    delegate_model: str | None = None
-    delegate_api_key: str | None = None
-    delegate_api_base: str | None = None
-
-    def to_dict(self) -> dict[str, Any]:
-        return asdict(self)
-
-    @classmethod
-    def from_raw(cls, raw: Any) -> SandboxLmRuntimeConfig:
-        if not isinstance(raw, dict):
-            raise ValueError("Sandbox LM config must be a dict.")
-        model = _normalize_optional_text(raw.get("model"))
-        api_key = _normalize_optional_text(raw.get("api_key"))
-        if model is None or api_key is None:
-            raise ValueError("Sandbox LM config requires model and api_key.")
-        return cls(
-            model=model,
-            api_key=api_key,
-            api_base=_normalize_optional_text(raw.get("api_base")),
-            max_tokens=_coerce_positive_int(raw.get("max_tokens")) or 64_000,
-            delegate_model=_normalize_optional_text(raw.get("delegate_model")),
-            delegate_api_key=_normalize_optional_text(raw.get("delegate_api_key")),
-            delegate_api_base=_normalize_optional_text(raw.get("delegate_api_base")),
-        )
-
-
-# ---------------------------------------------------------------------------
-# Section 3: Staged-context types
-# ---------------------------------------------------------------------------
-
-
-@dataclass(slots=True)
-class ContextSource:
-    """Host-sourced local context staged into the Daytona workspace."""
-
-    source_id: str
-    kind: str
-    host_path: str
-    staged_path: str
-    source_type: str | None = None
-    extraction_method: str | None = None
-    file_count: int = 1
-    skipped_count: int = 0
-    warnings: list[str] = field(default_factory=list)
-
-    @classmethod
-    def from_raw(cls, raw: Any) -> ContextSource:
-        if not isinstance(raw, dict):
-            raise ValueError("Context source payload must be a dict.")
-        source_id = _normalize_optional_text(raw.get("source_id"))
-        kind = _normalize_optional_text(raw.get("kind"))
-        host_path = _normalize_optional_text(raw.get("host_path"))
-        staged_path = _normalize_optional_text(raw.get("staged_path"))
-        if (
-            source_id is None
-            or kind is None
-            or host_path is None
-            or staged_path is None
-        ):
-            raise ValueError(
-                "Context source requires source_id, kind, host_path, and staged_path."
-            )
-        return cls(
-            source_id=source_id,
-            kind=kind,
-            host_path=host_path,
-            staged_path=staged_path,
-            source_type=_normalize_optional_text(raw.get("source_type")),
-            extraction_method=_normalize_optional_text(raw.get("extraction_method")),
-            file_count=_coerce_positive_int(raw.get("file_count")) or 1,
-            skipped_count=_coerce_nonnegative_int(raw.get("skipped_count")) or 0,
-            warnings=[
-                str(item)
-                for item in raw.get("warnings", []) or []
-                if item is not None and str(item).strip()
-            ],
-        )
-
-    def to_dict(self) -> dict[str, Any]:
-        return asdict(self)
 
 
 @dataclass(slots=True)
@@ -282,100 +374,89 @@ class DaytonaSmokeResult:
 
 
 # ---------------------------------------------------------------------------
-# Section 4: Chat/session normalization helpers (formerly state.py)
+# Sandbox spec-building helpers (formerly spec_runtime.py)
 # ---------------------------------------------------------------------------
 
-
-def render_final_text(value: Any) -> str:
-    if value is None:
-        return ""
-    if isinstance(value, str):
-        return value
-    if isinstance(value, dict):
-        for key in ("final_markdown", "summary", "text", "content", "message"):
-            candidate = value.get(key)
-            if isinstance(candidate, str) and candidate.strip():
-                return candidate
-        nested_value = value.get("value")
-        if nested_value is not value:
-            nested_text = render_final_text(nested_value)
-            if nested_text:
-                return nested_text
-    try:
-        return json.dumps(value, indent=2, ensure_ascii=False)
-    except (TypeError, ValueError):
-        return str(value)
+DEFAULT_SANDBOX_LABELS: dict[str, str] = {"managed-by": "fleet-rlm"}
 
 
-def dedupe_paths(paths: list[str]) -> list[str]:
-    seen: set[str] = set()
-    ordered: list[str] = []
-    for item in paths:
-        normalized = str(item or "").strip()
-        if not normalized or normalized in seen:
-            continue
-        seen.add(normalized)
-        ordered.append(normalized)
-    return ordered
+def default_sandbox_name(*, now: datetime.datetime | None = None) -> str:
+    """Return the dashboard-friendly default sandbox name."""
+    timestamp = now or datetime.datetime.now(datetime.timezone.utc)
+    return f"fleet-rlm-{timestamp:%Y%m%d-%H%M%S}"
 
 
-def history_messages(history: Any) -> list[dict[str, str]]:
-    messages = getattr(history, "messages", [])
-    if isinstance(messages, list):
-        return [item for item in messages if isinstance(item, dict)]
-    return []
+def merge_sandbox_labels(
+    *,
+    default_labels: Mapping[str, str] | None = None,
+    labels: Mapping[str, str] | None = None,
+) -> dict[str, str]:
+    """Merge runtime-default labels with caller-provided labels."""
+    merged = dict(default_labels or DEFAULT_SANDBOX_LABELS)
+    if labels:
+        merged.update(labels)
+    return merged
 
 
-def normalize_history_turn(raw: dict[str, Any]) -> dict[str, str] | None:
-    user_request = str(raw.get("user_request", "") or "").strip()
-    assistant_response = render_final_text(raw.get("assistant_response", "")).strip()
-    if not user_request and not assistant_response:
-        return None
-    return {
-        "user_request": user_request,
-        "assistant_response": assistant_response,
-    }
+def build_sandbox_spec(
+    *,
+    default_labels: Mapping[str, str] | None = None,
+    name: str | None = None,
+    volume_name: str | None = None,
+    volume_subpath: str | None = None,
+    image: Any = None,
+    snapshot: str | None = None,
+    env_vars: Mapping[str, str] | None = None,
+    labels: Mapping[str, str] | None = None,
+    cpu: int | None = None,
+    memory: int | None = None,
+    disk: int | None = None,
+    auto_stop_interval: int | None = 30,
+    auto_archive_interval: int | None = 60,
+    auto_delete_interval: int | None = None,
+    network_block_all: bool | None = None,
+    network_allow_list: str | None = None,
+) -> SandboxSpec:
+    """Build a ``SandboxSpec`` with Daytona runtime defaults applied."""
+    from .snapshot_runtime import resolve_default_snapshot
+    from .volume_runtime import DAYTONA_PERSISTENT_VOLUME_MOUNT_PATH
 
-
-def normalized_history_messages(history: Any) -> list[dict[str, str]]:
-    normalized: list[dict[str, str]] = []
-    for item in history_messages(history):
-        turn = normalize_history_turn(item)
-        if turn is not None:
-            normalized.append(turn)
-    return normalized
-
-
-def normalized_context_sources(raw: Any) -> list[ContextSource]:
-    if not isinstance(raw, list):
-        return []
-    normalized: list[ContextSource] = []
-    for item in raw:
-        try:
-            normalized.append(ContextSource.from_raw(item))
-        except Exception:
-            continue
-    return normalized
+    return SandboxSpec(
+        name=name or default_sandbox_name(),
+        language="python",
+        image=image,
+        snapshot=resolve_default_snapshot(image=image, snapshot=snapshot),
+        volume_name=volume_name,
+        volume_mount_path=str(DAYTONA_PERSISTENT_VOLUME_MOUNT_PATH),
+        volume_subpath=volume_subpath,
+        env_vars=dict(env_vars) if env_vars else None,
+        labels=merge_sandbox_labels(default_labels=default_labels, labels=labels),
+        ephemeral=True,
+        auto_stop_interval=auto_stop_interval,
+        auto_archive_interval=auto_archive_interval,
+        auto_delete_interval=auto_delete_interval,
+        cpu=cpu,
+        memory=memory,
+        disk=disk,
+        network_block_all=network_block_all,
+        network_allow_list=network_allow_list,
+    )
 
 
 __all__ = [
-    # Serialization helpers
-    "_normalize_optional_text",
-    "_coerce_positive_int",
-    "_coerce_nonnegative_int",
-    # Sandbox spec and LM types
-    "DaytonaRunCancelled",
-    "SandboxSpec",
-    "SandboxLmRuntimeConfig",
-    # Context types
     "ContextSource",
-    # Smoke result
+    "DEFAULT_SANDBOX_LABELS",
+    "DaytonaRunCancelled",
     "DaytonaSmokeResult",
-    # State helpers (formerly state.py)
-    "render_final_text",
+    "SandboxLmRuntimeConfig",
+    "SandboxSpec",
+    "build_sandbox_spec",
     "dedupe_paths",
+    "default_sandbox_name",
     "history_messages",
+    "merge_sandbox_labels",
     "normalize_history_turn",
-    "normalized_history_messages",
     "normalized_context_sources",
+    "normalized_history_messages",
+    "render_final_text",
 ]
