+++ tests/unit/integrations/daytona/test_interpreter.py
@@ -324,6 +324,32 @@ def test_daytona_interpreter_execute_direct_reuses_context_and_returns_final_out
     assert len(runtime.session.sandbox.code_interpreter.contexts) == 1
 
 
+def test_daytona_interpreter_strips_trailing_dspy_completed_marker() -> None:
+    runtime = _FakeRuntime()
+    interpreter = DaytonaInterpreter(runtime=runtime)
+
+    result = interpreter.execute("SUBMIT(answer='ok')\n]] [[ ## completed ## ]]")
+
+    assert isinstance(result, FinalOutput)
+    assert getattr(result, "output") == {"answer": "ok"}
+    executed = runtime.session.sandbox.code_interpreter.run_calls[-1]
+    assert "[[ ## completed ## ]]" not in executed
+    assert "]]" not in executed
+
+
+def test_daytona_interpreter_returns_structured_prepare_error_for_bad_marker_leak() -> (
+    None
+):
+    runtime = _FakeRuntime()
+    interpreter = DaytonaInterpreter(runtime=runtime)
+
+    result = interpreter.execute("if True:\n    pass\nelse\n[[ ## completed ## ]]")
+
+    payload = json.loads(str(result))
+    assert payload["status"] == "error"
+    assert payload["reason"] == "code_prepare_error"
+
+
 def test_daytona_interpreter_uses_bridge_for_llm_queries(monkeypatch) -> None:
     runtime = _FakeRuntime()
     interpreter = DaytonaInterpreter(runtime=runtime)
@@ -376,6 +402,95 @@ def test_daytona_interpreter_uses_bridge_for_llm_queries(monkeypatch) -> None:
     assert isinstance(result, FinalOutput)
     assert getattr(result, "output") == {"answer": "HOST:hello"}
     assert "llm_query" in captured["tools"]
+
+
+def test_daytona_interpreter_bridge_detection_uses_cleaned_code(monkeypatch) -> None:
+    runtime = _FakeRuntime()
+    interpreter = DaytonaInterpreter(runtime=runtime)
+    interpreter.llm_query = lambda prompt: f"HOST:{prompt}"  # type: ignore[method-assign]
+
+    captured: dict[str, Any] = {}
+
+    class _FakeBridge:
+        def __init__(self, *, sandbox: Any, context: Any) -> None:
+            captured["sandbox"] = sandbox
+            captured["context"] = context
+
+        def bind_context(self, context: Any) -> None:
+            captured["bound_context"] = context
+
+        async def async_tools(self, tools: dict[str, Any]) -> None:
+            captured["tools"] = dict(tools)
+
+        async def aexecute(
+            self,
+            *,
+            code: str,
+            timeout: int,
+            tool_executor,
+            on_stdout=None,
+            on_stderr=None,
+        ):
+            del timeout, on_stdout, on_stderr
+            captured["code"] = code
+            answer = tool_executor("llm_query", ["hello"], {})
+            payload = f"{_FINAL_OUTPUT_MARKER}{json.dumps({'answer': answer})}{_FINAL_OUTPUT_MARKER}"
+            return DaytonaBridgeExecution(
+                result=_FakeExecutionResult(),
+                stdout=payload,
+                stderr="",
+                callback_count=1,
+            )
+
+        async def aclose(self) -> None:
+            captured["closed"] = True
+
+    monkeypatch.setattr(
+        "fleet_rlm.integrations.daytona.interpreter.DaytonaToolBridge",
+        _FakeBridge,
+    )
+
+    result = interpreter.execute(
+        "Code:\n```python\nanswer = llm_query('hello')\nSUBMIT(answer=answer)\n```\n[[ ## completed ## ]]"
+    )
+
+    assert isinstance(result, FinalOutput)
+    assert getattr(result, "output") == {"answer": "HOST:hello"}
+    assert "[[ ## completed ## ]]" not in captured["code"]
+    assert "```" not in captured["code"]
+
+
+def test_daytona_interpreter_bridge_injection_error_is_structured(monkeypatch) -> None:
+    runtime = _FakeRuntime()
+    interpreter = DaytonaInterpreter(runtime=runtime)
+
+    class _FakeBridge:
+        def __init__(self, *, sandbox: Any, context: Any) -> None:
+            del sandbox, context
+
+        def bind_context(self, context: Any) -> None:
+            del context
+
+        async def async_tools(self, tools: dict[str, Any]) -> None:
+            del tools
+            raise CodeInterpreterError(
+                "Failed to inject tool 'store_evidence': invalid syntax"
+            )
+
+        async def aclose(self) -> None:
+            pass
+
+    monkeypatch.setattr(
+        "fleet_rlm.integrations.daytona.interpreter.DaytonaToolBridge",
+        _FakeBridge,
+    )
+
+    result = interpreter.execute("answer = llm_query('hello')\nSUBMIT(answer=answer)")
+
+    payload = json.loads(str(result))
+    assert payload["status"] == "error"
+    assert payload["reason"] == "tool_error"
+    assert "store_evidence" in payload["error"]
 
 
 def test_daytona_interpreter_exports_context_id_for_resume() -> None:
